Remove conditional import logic of onnxruntime
